requirements.txt
django == 5.1.2
djangorestframework == 3.15.2
pytest-django == 4.9.0
model-bakery == 1.20.0